refactor(ui): split vulns page into queue and drawer components
Extract findings table and evidence drawer from /vulns, sync filters to URL query params, and keep the page test green with navigation mocks.
